More locations added to logic
Granny's blue potion, and rupee rewards for selling items were added to logic, but these locations are not yet randomized. This also puts poes in logic, but poe locations are not included yet. Also included logic for a potential setting to re-enable like likes eating tunics but this is not added either.